Rename openshift base dir to pkg
