+++ pkg/controller/die.go
