+++ pkg/operator/filter.go
@@ -1,6 +1,6 @@
 package operator
 
-import "github.com/openshift/operator-boilerplate-legacy/openshift/controller"
+import "github.com/openshift/operator-boilerplate-legacy/pkg/controller"
 
 func FilterByNames(names ...string) controller.Filter {
 	return controller.FilterByNames(nil, names...)
