+++ pkg/operator/option_test.go
